# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Luna AI API",
    "version": "1.0.0",
    "description": "API Documentation for Luna AI Streamer - AI-powered streamer with VTube Studio integration, emotion system, and Rock Paper Scissors game",
    "contact": {
      "name": "Luna AI Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local development server"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/luna/admin/clear-memory": {
      "get": {
        "tags": ["Admin"],
        "summary": "Clear user memory",
        "description": "Clear memory for a specific user or all users (Admin only)",
        "security": [
          {
            "AdminSecret": []
          }
        ],
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Username (optional, clears all if not provided)"
          },
          {
            "name": "secret",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Admin secret (alternative to header)"
          }
        ],
        "responses": {
          "200": {
            "description": "Memory cleared",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/luna/admin/reset-personality": {
      "get": {
        "tags": ["Admin"],
        "summary": "Reset personality",
        "description": "Reset Luna's personality to default (Admin only)",
        "security": [
          {
            "AdminSecret": []
          }
        ],
        "responses": {
          "200": {
            "description": "Personality reset",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/luna/admin/clear-cache": {
      "get": {
        "tags": ["Admin"],
        "summary": "Clear response cache",
        "description": "Clear all cached responses (Admin only)",
        "security": [
          {
            "AdminSecret": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cache cleared",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/luna/admin/errors": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get error logs",
        "description": "Get recent error logs (Admin only)",
        "security": [
          {
            "AdminSecret": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Maximum number of errors to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Error logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/luna/admin/reset-stats": {
      "get": {
        "tags": ["Admin"],
        "summary": "Reset statistics",
        "description": "Reset all system statistics (Admin only)",
        "security": [
          {
            "AdminSecret": []
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics reset",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Invalid request"
          },
          "message": {
            "type": "string",
            "example": "Error message"
          }
        }
      },
...
    },
    "securitySchemes": {
      "AdminSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-secret",
        "description": "Admin secret key (set ADMIN_SECRET in .env)"
      },
      "PurchaseSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-purchase-secret",
        "description": "Purchase webhook secret (set PURCHASE_SECRET in .env)"
      }
    }
  }
}